Refactor: Move quota evaluation to dedicated quotas package
Move QuotaLimits and quota evaluation logic from controller package to
a new apps/provisioning/pkg/quotas package. This improves separation of
concerns and makes the quota logic reusable.

Changes:
- Create apps/provisioning/pkg/quotas package with QuotaLimits struct
- Move BuildQuotaCondition to QuotaLimits.EvaluateCondition method
- Move quota unit tests to quotas package
- Update sync worker to use new quotas package
- Add comment explaining why quotas are evaluated in sync worker
  (pull) rather than controller: sync worker performs reconciliation,
  cleans up resources, and updates stats, making it the natural place
  to evaluate quotas against actual resource state.
