+++ pkg/registry/apis/provisioning/jobs/sync/worker.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/grafana/grafana-app-sdk/logging"
 	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
+	"github.com/grafana/grafana/apps/provisioning/pkg/quotas"
 	"github.com/grafana/grafana/apps/provisioning/pkg/repository"
 	"github.com/grafana/grafana/pkg/infra/tracing"
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/controller"
@@ -215,11 +216,18 @@ func (r *SyncWorker) Process(ctx context.Context, repo repository.Repository, jo
 		finalSpan.SetAttributes(attribute.Int("stats.unexpected_count", len(stats.Managed)))
 	}
 
-	// Update Quota condition based on stats and configured limits
-	quotaLimits := controller.QuotaLimits{
+	// Update Quota condition based on stats and configured limits.
+	// Quotas are evaluated here in the sync worker (pull) rather than in the controller because:
+	// 1. The sync worker performs reconciliation and eventually cleans up resources, so it has
+	//    the most up-to-date view of what resources actually exist.
+	// 2. The sync worker is responsible for updating stats after each sync operation, making it
+	//    the natural place to evaluate quotas against those stats.
+	// 3. This ensures quota conditions reflect the actual resource state after reconciliation,
+	//    not just what the controller thinks should exist.
+	quotaLimits := quotas.QuotaLimits{
 		MaxResources: r.maxResourcesPerRepository,
 	}
-	quotaCondition := controller.BuildQuotaCondition(repoStats, quotaLimits)
+	quotaCondition := quotaLimits.EvaluateCondition(repoStats)
 	if quotaConditionOps := controller.BuildConditionPatchOpsFromExisting(cfg.Status.Conditions, cfg.GetGeneration(), quotaCondition); quotaConditionOps != nil {
 		patchOperations = append(patchOperations, quotaConditionOps...)
 	}
